fixing: #1611 using drawer pattern
